requirements.txt
langchain
langchain-openai
openai
beautifulsoup4
pydantic
python-dotenv
requests